Moved configuration of Eventbus => spring-vaadin-eventbus VaadinEventbusConfiguration
